feat: Add NumberInputComponent.kt
