+++ app/src/main/java/com/example/kala/screens/components/buttons/LargeButtonComponent.kt
@@ -1,4 +1,4 @@
-package com.example.kala.screens.components
+package com.example.kala.screens.components.buttons
 
 import android.annotation.SuppressLint
 import android.os.Build
@@ -8,10 +8,11 @@ import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
@@ -36,19 +37,18 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.kala.configuration.MediumButtonConfiguration
+import com.example.kala.configuration.LargeButtonConfiguration
 import com.example.kala.configuration.SVG_DESCRIPTION
 import com.example.kala.configuration.invalidArgument
 import com.example.kala.ui.theme.BoneWhite
 
-
 @RequiresApi(Build.VERSION_CODES.N)
 @Composable
-fun MediumButton(configuration: MediumButtonConfiguration) {
+fun LargeButton(configuration: LargeButtonConfiguration) {
     Box(
         modifier = Modifier
-            .height(150.dp)
-            .width(130.dp)
+            .height(60.dp)
+            .width(300.dp)
             .shadow(10.dp, shape = RoundedCornerShape(10.dp))
     ) {
         Button(
@@ -57,25 +57,34 @@ fun MediumButton(configuration: MediumButtonConfiguration) {
                 invalidArgument()
             },
             modifier = Modifier
-                .height(150.dp)
-                .width(130.dp)
+                .height(60.dp)
+                .width(300.dp)
                 .border(2.dp, Color.Black, RoundedCornerShape(10.dp)),
             colors = ButtonDefaults.buttonColors(Color.White),
             shape = RoundedCornerShape(10.dp),
             contentPadding = PaddingValues(10.dp),
-        ) {
-            Column(
-                modifier = Modifier
-                    .fillMaxSize(),
-                verticalArrangement = Arrangement.Center,
-                horizontalAlignment = Alignment.CenterHorizontally
+
             ) {
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth(),
+                horizontalArrangement = Arrangement.SpaceBetween,
+                verticalAlignment = Alignment.CenterVertically,
+            ) {
+                Text(
+                    text = configuration.toString(),
+                    color = Color.Black,
+                    fontSize = 28.sp,
+                    textAlign = TextAlign.Justify,
+                    fontWeight = FontWeight.Bold,
+                    modifier = Modifier.weight(1f),
+                )
+
                 Box(
                     modifier = Modifier
                         .border(2.dp, Color.Black, shape = CircleShape)
-                        .padding(8.dp)
                         .clip(CircleShape)
-                        .size(50.dp)
+                        .size(40.dp)
                         .clip(CircleShape)
                         .padding(7.dp),
                     contentAlignment = Alignment.Center,
@@ -87,23 +96,18 @@ fun MediumButton(configuration: MediumButtonConfiguration) {
                         contentDescription = SVG_DESCRIPTION
                     )
                 }
-                Text(
-                    text = configuration.toString(),
-                    color = Color.Black,
-                    fontSize = 25.sp,
-                    textAlign = TextAlign.Center,
-                    fontWeight = FontWeight.Bold,
-                )
+
             }
         }
     }
+
 }
 
 @SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
 @RequiresApi(Build.VERSION_CODES.N)
 @Preview
 @Composable
-fun PreviewMediumButton() {
+fun LargeButtonPreview() {
     Scaffold {
         LazyColumn (
             modifier = Modifier
@@ -112,12 +116,13 @@ fun PreviewMediumButton() {
             horizontalAlignment = Alignment.CenterHorizontally,
             verticalArrangement = Arrangement.Center,
         ){
-            items(MediumButtonConfiguration.entries.toTypedArray()){
-                    value ->
-                MediumButton(configuration = value)
+            items(LargeButtonConfiguration.entries.toTypedArray()){
+                value ->
+                    LargeButton(configuration = value)
                 Spacer(modifier = Modifier.padding(5.dp))
 
             }
         }
     }
+
 }
